Extract Prompt Studio advanced highlight helpers
